+++ BALSAMIC/constants/constants.py
@@ -1,7 +1,6 @@
 """General use constants."""
+from enum import StrEnum
 from typing import List
-
-from BALSAMIC.utils.class_types import StrEnum
 
 EXIT_SUCCESS: int = 0
 EXIT_FAIL: int = 1
@@ -17,7 +16,7 @@ class LogLevel(StrEnum):
     CRITICAL = "CRITICAL"
 
 
-LOG_LEVELS: List[LogLevel] = [level.value for level in LogLevel]
+LOG_LEVELS: List[LogLevel] = [level for level in LogLevel]
 
 
 class FileType(StrEnum):