+++ CHANGELOG.rst
@@ -25,7 +25,9 @@ Added:
 * `PureCN` container https://github.com/Clinical-Genomics/BALSAMIC/pull/1255
 * `GATK` container https://github.com/Clinical-Genomics/BALSAMIC/pull/1266
 * Resolved FASTQ paths to sample dictionary (balsamic logging) https://github.com/Clinical-Genomics/BALSAMIC/pull/1275
-* `LOH` to TGA workflow https://Clinical-Genomics/BALSAMIC/pull/1278
+* `LOH` and CNVs from PureCN to TGA workflow https://Clinical-Genomics/BALSAMIC/pull/1278
+* Command-line arguments and rules for creation of GENS files https://github.com/Clinical-Genomics/BALSAMIC/pull/1279
+
 
 Changed:
 ^^^^^^^^
@@ -54,6 +56,9 @@ Changed:
 * Migrate Metrics models to pydantic v2 https://github.com/Clinical-Genomics/BALSAMIC/pull/1270
 * Migrate Snakemake models to pydantic v2 https://github.com/Clinical-Genomics/BALSAMIC/pull/1268
 * Migrate Cache models to pydantic v2 https://github.com/Clinical-Genomics/BALSAMIC/pull/1274
+* Made BALSAMIC compatible with multiple PON creation workflows https://github.com/Clinical-Genomics/BALSAMIC/pull/1279
+* Use StrEnum from python enum https://github.com/Clinical-Genomics/BALSAMIC/pull/1303
+
 
 Fixed:
 ^^^^^^